Migrate to typescript and base
