--- packages/mui-material-next/src/ButtonBase/index.d.ts
@@ -1,8 +1,0 @@
-export { default } from './ButtonBase';
-export * from './ButtonBase';
-
-export { default as buttonBaseClasses } from './buttonBaseClasses';
-export * from './buttonBaseClasses';
-
-export { default as touchRippleClasses } from './touchRippleClasses';
-export * from './touchRippleClasses';
