+++ packages/mui-material-next/src/ButtonBase/index.ts
@@ -3,6 +3,3 @@ export { default } from './ButtonBase';
 
 export { default as buttonBaseClasses } from './buttonBaseClasses';
 export * from './buttonBaseClasses';
-
-export { default as touchRippleClasses } from './touchRippleClasses';
-export * from './touchRippleClasses';
